filehandling

+++ SourceCode/SRI/Main/GetAndCompareLogs.psm1
@@ -59,19 +59,30 @@ Function WriteXMLElement([System.XMl.XmlTextWriter] $XmlWriter, [String] $startE
 function ImportCompareExport($importPath, $exportPath){
     $eventLogIdsToCheck = (6, 21, 24, 102, 104, 106, 129, 169, 200, 201, 4624, 4634, 4648, 4656, 4658, 4660, 4661, 4663, 4672, 4673, 4688, 4689, 4690, 4720, 4726, 4728, 4729, 4768,4769, 4946, 5140, 5142, 5144, 5145, 5154, 5156, 7036, 7045, 8222, 20001)
     $appAndServIdsToCheck = (106, 200, 129, 201, 102, 6, 169, 21, 24)
+    $isCurrentPath = $true
 
     $resultXML = $PSScriptRoot + "\resultOfEventLogs.xml"
     if ($exportPath) {
         $resultXML = $exportPath + "\resultOfEventLogs.xml"
     }
+
     $importEventLogs = $PSScriptRoot + "\myeventlogs.csv"
     $importAppAndServLogs = $PSScriptRoot + "\myappandservlogs.csv"
     if ($importPath) {
+        $isCurrentPath = $false
         $importEventLogs = $importPath + "\myeventlogs.csv"
         $importAppAndServLogs = $importPath + "\myappandservlogs.csv"
+        if (-not [System.IO.File]::Exists($importEventLogs)) {
+            Write-Host "File $importEventLogs does not exist!" -ForegroundColor Red
+            return
+        }
+        if (-not [System.IO.File]::Exists($importAppAndServLogs)) {
+            Write-Host "File $importAppAndServLogs does not exist!" -ForegroundColor Red
+            return
+        }        
     }
+
     $xmlWriter = New-Object System.XMl.XmlTextWriter($resultXML,$Null) 
-    
     $myEventLogs = Import-Csv $importEventLogs -Encoding UTF8
 
     Write-Host "Comparing Found EventLogs to Checklist"
@@ -114,8 +125,10 @@ function ImportCompareExport($importPath, $exportPath){
     
     Write-Host "Done Event Logs!!!"
 
-    Remove-Item $importEventLogs
-    Remove-Item $importAppAndServLogs
+    if ($isCurrentPath) {
+        Remove-Item $importEventLogs
+        Remove-Item $importAppAndServLogs
+    }
 }
 
 Export-ModuleMember -Function GetEventLogsAndExport, ImportCompareExport, GetApplicationAndServiceLogs
